Change NewsFeed Post Modal to Scene
