+++ App/Navigation/routes.js
@@ -2,6 +2,7 @@ import Home from '~/Scene/Home';
 import Login from '~/Scene/Auth/Login';
 import ForgotPassword from '~/Scene/Auth/ForgotPassword/';
 import NewsFeed from '~/Scene/NewsFeed';
+import NewsFeedPosting from '~/Scene/NewsFeed/NewsFeedPosting/';
 import Notification from '~/Scene/Notification';
 import People from '~/Scene/People';
 import Profile from '~/Scene/Profile';
@@ -65,6 +66,16 @@ export const ROUTES = {
       type: 'material-community',
     },
     initial: true,
+  },
+  newsFeedPosting: {
+    name: 'New Post',
+    path: '/news-feed/post',
+    screen: NewsFeedPosting,
+    icon: {
+      name: 'newspaper',
+      type: 'material-community',
+    },
+    // initial: true,
   },
   myAgenda: {
     name: 'My Agenda',
